% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_previous_results_analysis.R
\name{AnalysisPreviousUI}
\alias{AnalysisPreviousUI}
\title{Title}
\usage{
AnalysisPreviousUI(id, dtIndikatorensettings, all_members = FALSE)
}
\arguments{
\item{all_members}{}
}
\description{
Title
}
