% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_reactive_helper_functions.R
\name{calculatecenterfunc}
\alias{calculatecenterfunc}
\title{Title}
\usage{
calculatecenterfunc(centervalue, x, offset = 0, trim = 0,
  na.rm = TRUE)
}
\arguments{
\item{centervalue}{one of mean, min, max, median, q1, q3; numeric value; NA}

\item{trim}{works only for mean}

\item{na.rm}{}
}
\value{
defaults to mean
}
\description{
Title
}
\examples{

}
