% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{crossjoinFunc}
\alias{crossjoinFunc}
\title{Crossjoin of two data.tables}
\usage{
crossjoinFunc(datatable_x1, datatable_x2)
}
\arguments{
\item{x1}{first data.table}

\item{x2}{second data.table}
}
\value{
crossjoin of data.tables x1 and x2
}
\description{
Crossjoin of two data.tables
}
\examples{

}
