% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_input_functions.R
\name{getIndikatorensetting}
\alias{getIndikatorensetting}
\title{Get Settings of Indicators}
\usage{
getIndikatorensetting(x, util_func = "prop", util_mean = "mean",
  util_offset = ifelse(util_func == "antiprop", 10, 0), util_scale = 100,
  agg_func = "weighted.sum", include_parent = TRUE, open.maxdepth = Inf,
  standardweight = 30)
}
\arguments{
\item{x}{Configuration list, see \code{\link{validateConfig}}.}

\item{util_func}{standard utitily function.}

\item{util_mean}{standard utility mean.}

\item{util_offset}{standard utility offset.}

\item{util_scale}{standard utility scale.}

\item{agg_func}{standard aggregation function.}

\item{include_parent}{Should new settings apply also to parent itself, or only to children?}

\item{open.maxdepth}{}
}
\description{
Get Settings of Indicators
}
