% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{getOpen.Maxdepth}
\alias{getOpen.Maxdepth}
\title{Get minimum of open.maxdepth and x$open.maxdepth}
\usage{
getOpen.Maxdepth(x, open.maxdepth = Inf)
}
\arguments{
\item{x}{fgh}

\item{open.maxdepth}{dfgh}
}
\value{
returns min( x$open.maxdepth, open.maxdepth) if x$open.maxdepth exists, if not open.maxdepth
}
\description{
Compares open.maxdepth and x$open.maxdepth and returns lower one, if present.
}
\examples{
getOpen.Maxdepth( NULL, 12)
getOpen.Maxdepth( list(open.maxdepth=10), 12)

}
