% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{getwidth}
\alias{getwidth}
\title{Even spread of difference}
\usage{
getwidth(x1, x2, share = 20)
}
\arguments{
\item{x1}{lower value}

\item{x2}{higher value}

\item{share}{number of partitions}
}
\value{
width of each partition
}
\description{
Even spread of difference
}
\examples{
getwidth(1,10,9)
getwidth(1,10,4)

}
