% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{\%||\%}
\alias{\%||\%}
\title{Fallback for Null-Value.}
\usage{
a \%||\% b
}
\arguments{
\item{a}{First value.}

\item{b}{Fallback.}
}
\value{
if (!is.null(a)) a else b
}
\description{
Useful for example for reinitialising inputs.
}
\note{
see \url{https://stackoverflow.com/questions/33137546/switch-between-layouts-reactively-with-shiny?noredirect=1&lq=1}.
}
