% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastorage_functions.R
\name{gs_sheetexists}
\alias{gs_sheetexists}
\title{Testing if googlesheet exists in open Googlesheet-location.}
\usage{
gs_sheetexists(place)
}
\arguments{
\item{place}{Sheet to be tested for existence.}
}
\value{
True if sheet exists, False otherwise.
}
\description{
Testing if googlesheet exists in open Googlesheet-location.
}
\note{
Datastorage must have been initialised before using
      \code{\link{initialize_datastorage()}}.
}
\examples{

}
