% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_functions.R
\name{indentedRow}
\alias{indentedRow}
\title{Creation of indented fluidrows}
\usage{
indentedRow(..., indention = 0, tagListWrap = FALSE)
}
\arguments{
\item{...}{Elements to be included (each on several row).}

\item{indention}{Integer between 0 and 11. Indention.}

\item{tagListWrap}{Should fluidrow be wrapped into \code{\link{shiny:tagList}}.}
}
\value{
\code{\link{shiny:fluidRow}}/\code{\link{shiny:column}} with indented elements.
}
\description{
Creates an indented fluidrow
}
