% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastorage_functions.R
\name{initialize_datastorage}
\alias{initialize_datastorage}
\title{Initializes external data storage.}
\usage{
initialize_datastorage(data, method, place)
}
\arguments{
\item{place}{}
}
\description{
Initializes external data storage.
}
