% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdsupportshiny.R
\docType{package}
\name{mcdsupportshiny}
\alias{mcdsupportshiny}
\alias{mcdsupportshiny-package}
\title{mcdsupportshiny: Customisable graphical application to support for multi-criteria decision problems using shiny.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
Most essentially, however, is adapting it to your problem - your data and
your categories - and to deploy it online .
}
\section{Configuration}{
 Adapting it to your problem
To adapt it to your ...
}

\section{Deploying App online}{
NA
}

\section{Foo functions}{

The foo functions ...
}

