% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_functions.R
\name{rColorVector}
\alias{rColorVector}
\alias{rColorSliders}
\title{Recursively turn Configurations list into a vector of colors}
\usage{
rColorVector(x, color = "", color_parent = TRUE)

rColorSliders(x, id = NULL, slidername = "sl", color_parent = TRUE,
  collapse = "\\n")
}
\arguments{
\item{x}{Configuration list, see \code{\link{validateConfig}}.}

\item{color}{Default color.}

\item{color_parent}{Should parent itself also be colored in new color or only the children?}

\item{collapse}{should all styles be collapsed into a single style-tag (default), or not (collapse=NULL)}
}
\value{
named vector of colors, corresponding to flatted list. names: element names.
}
\description{
Recursively turn Configurations list into a vector of colors
}
\section{Functions}{
\itemize{
\item \code{rColorSliders}: Vector of colors converted into a vector of CSS-Styles to use for ION.rangeSlider.
}}

\note{
See \url{https://stackoverflow.com/questions/36906265/how-to-color-sliderbar-sliderinput.}
}
\examples{

}
