% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_functions.R
\name{recColorVector}
\alias{recColorVector}
\title{Internal Version of rColorVector}
\usage{
recColorVector(x, num = 0, color = "", color_parent = TRUE)
}
\arguments{
\item{num}{Position in flatted List, because recursive list is processed iteratively}

\item{color_parent}{}
}
\description{
Internal Version of rColorVector
}
