% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_output_functions.R
\name{recursiveTrue}
\alias{recursiveTrue}
\title{recursive indicating wether element or one of parents is false}
\usage{
recursiveTrue(names, states, parents)
}
\arguments{
\item{names}{vector of names (character or integer)}

\item{states}{boolean vector}

\item{parents}{vector of parents (character or integer), only first one is used, if several parents}
}
\value{
named vector, ordered by data$names, of logical indicating if one of the parents`state or state itself is false
}
\description{
recursive indicating wether element or one of parents is false
}
\examples{

}
