% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_input_functions.R
\name{rgetIndikatorensetting}
\alias{rgetIndikatorensetting}
\title{Internal version of getIndikatorensetting}
\usage{
rgetIndikatorensetting(x, depth = 0, parent = "Szenarioergebnis",
  positive_utility_settings = utility_settings(),
  negative_utility_settings = utility_settings(util_func = "negprop"),
  agg_func = "weighted.sum", include_parent = TRUE, open.maxdepth = Inf,
  color = NA, minweight = 0, standardweight = 30)
}
\arguments{
\item{positive_utility_settings}{standard utitily function settings, \code{\link{utility_settings}()}.}

\item{negative_utility_settings}{standard utitily function settings for negative weights, \code{\link{utility_settings}(util_func="negprop")} .}

\item{color}{}

\item{minweight}{only to determine, if negative settings should be included, see \code{\link{rSliderGuiInput}}.}
}
\value{

}
\description{
Internal version of getIndikatorensetting
}
