% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_functions.R
\name{rSliderGuiInput}
\alias{rSliderGuiInput}
\alias{rSliderGui}
\title{Creates List of Sliders.}
\usage{
rSliderGuiInput(id, x, breaking = NULL, beschreibungs_text = NULL,
  title_text = "Please adjust the sliders according to your preferences.",
  mainpageposition = c("first", "last", "none"),
  parents_name = "Hauptkategorien", minweight = 0, maxweight = 100,
  standardweight = 30, open.maxdepth = Inf, cb_title = "I don't know",
  sliderlabel = "")

rSliderGui(input, output, session, slCbNames, dtBscCombinations)
}
\arguments{
\item{x}{Configuration list, see \code{\link{validateConfig}}.}

\item{breaking}{Should SliderGui be split up into a list of several pages, or deliver
one single list of sliders?
\describe{
   \item{NULL or FALSE}{No pagination.}
   \item{TRUE}{Put all subordinate elements of first order
               on extra pages, add main page containing only first order elements.}
   \item{Integer}{Put all subordinate elements of nth order on extra pages, add
                  main page containing elements up to nth order.
                  1 corresponds to TRUE.}
}
Maybe in future versions possibilities for putting special Elements on
extra pages by name will be added, right now this is not possible.}

\item{beschreibungs_text}{}

\item{title_text}{Which paragraph should be added as first element of each page?}

\item{mainpageposition}{One of "first", "last", "none".
Should main page be first or last or be omitted?}

\item{parents_name}{Label used to name root when creating fold-out-panel for root.}

\item{minweight}{Standard minimum slider weight.}

\item{maxweight}{Standard maximum slider weight.}

\item{standardweight}{Standard slider weight.}

\item{open.maxdepth}{From which depth on to create fold-out-panels, if not declared otherwise in x.
Maximum:12 (see \code{\link{shiny:column}}).Must be larger than breaking (if provided).}

\item{cb_title}{Default description label for checkbox.}

\item{sliderlabel}{Default class of slidercheckboxes, to use with \href{https://stackoverflow.com/a/42887905}{custom
prettify-function to have individual labels}.}
}
\value{
Slidergui: List of Sliders (breaking== NULL or breaking==FALSE)
                   or list of SliderGui-pages (breaking==TRUE or Breaking>0).
}
\description{
Creates List of Sliders.
}
\section{Functions}{
\itemize{
\item \code{rSliderGui}: list of two: sliderCheckBoxValues and collapsePanelValues
}}

