% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{setSliderGuiAttribs}
\alias{setSliderGuiAttribs}
\title{Setting Attributes of Elements within sliderGrui}
\usage{
setSliderGuiAttribs(x, element_name, depth, parent_name)
}
\arguments{
\item{x}{object}

\item{element_name}{value for element_name-attribute}

\item{depth}{value for depth-attribute}

\item{parent_name}{value for parent_name-attribute}
}
\value{
x with attributes element_name, depth, and parent_name
}
\description{
Returns x with attributes element_name, depth, and parent_name set to values
}
\note{
Setting attribute "names" (with \code{\link{stats:setNames}}) produces an errors, therefore
      attribute "element_name" is used.
}
