% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_slidercheckbox.R
\name{sliderCheckboxInput}
\alias{sliderCheckboxInput}
\alias{sliderCheckbox}
\alias{updateSliderCheckboxInput}
\title{Module sliderCheckbox: Slider with Checkbox}
\usage{
sliderCheckboxInput(id, description = "", min = 0, max = 100,
  value = 30, default = value, cb_title = "I don't know",
  sliderclass = "", width = "100\%")

sliderCheckbox(input, output, session, oldvalue = NULL)

updateSliderCheckboxInput(session, id, description = NULL, min = NULL,
  max = NULL, default = NULL, cb_title = NULL)
}
\arguments{
\item{id}{Modules id.}

\item{description}{Label for Slider.}

\item{min}{Slider minimum value.}

\item{max}{Slider maximum value.}

\item{value}{Slider initial value. If NA, Checkbox will be activated and default returned.}

\item{default}{Value to be returned if Checkbox is active.}

\item{cb_title}{Description Label for Checkbox.}

\item{width}{The width of the \code{\link[shiny]{sliderInput}}, e.g. '400px', or '100\%'; see \code{\link[shiny]{validateCssUnit}}.}

\item{session}{Passing the \code{session} object to function which was given to \code{shinyServer}.}

\item{oldvalue}{Value to be stored if value is NA. SLider will be set to oldvalue after Checkbox is
deactivated.}
}
\value{
fluidrow with slider and checkbox for Input, values as described below.
}
\description{
Creates a Slider with a Checkbox to mark "I don't know" and corresponding functionality.
}
\section{Functions}{
\itemize{
\item \code{sliderCheckbox}: Returns list of reactives and functions.
#TODO: Documentation.
Slidervalue if checkbox is FALSE, else default.
Value which sider had when checkbox was last checked will be returned in attribute "oldvalue".

\item \code{updateSliderCheckboxInput}: Updates static parts ofsliderCheckboxInput (not values!), ignoring all NULL-values.
}}

\note{
For Usage of Modules, see \url{https://shiny.rstudio.com/articles/modules.html}.
      To use the functionality, \code{\link[shinyjs]{shinyjs}} must have been loaded and
      \code{\link[shinyjs:useShinyjs]{initialised}}.
}
