% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_functions.R
\name{syncSliderGuiInputs}
\alias{syncSliderGuiInputs}
\title{Setting one SliderGuiInput to the state of another}
\usage{
syncSliderGuiInputs(sliderGui, oldSliderGui)
}
\arguments{
\item{sliderGui}{The sliderGui to be synched with the other.}

\item{oldSliderGui}{the old slider Gui (list object, without Parenthesis.)}
}
\description{
Setting one SliderGuiInput to the state of another
}
