% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_input_functions.R
\name{utility_settings}
\alias{utility_settings}
\title{Helper function. Standard configuration of utility function}
\usage{
utility_settings(util_func = "prop", util_fit_x1 = NA, y1 = NA,
  util_fit_x2 = "mean", y2 = 100)
}
\arguments{
\item{y1}{}
}
\value{
list of five:
list(util_func=util_func,
util_fit_x1=util_fit_x1,
y1=y1,
util_fit_x2=util_fit_x2,
y2=y2)
}
\description{
Helper function. Standard configuration of utility function
}
