% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_reactive_helper_functions.R
\name{utilityfunc}
\alias{utilityfunc}
\title{Convert Values of Indicators to utility-values.}
\usage{
utilityfunc(x, type, x1 = 0, y1 = switch(as.character(type), prop = 0,
  negprop = 200, antiprop = 200, 0), x2 = calculatecenterfunc("mean", x,
  offset = 0), y2 = 100, replace_NA = TRUE)
}
\arguments{
\item{x}{Numeric. Value to be converted.}

\item{type}{One of c("prop", "negprop", antiprop", "identity")}

\item{x1}{if (replace_NA=TRUE & is.na(x1)): 0}

\item{y1}{if (replace_NA=TRUE & is.na(y1)): switch(as.character(type),prop=0,negprop=200,antiprop=200,0)}

\item{x2}{if (replace_NA=TRUE & is.na(x2)): calculatecenterfunc("mean", x,offset=0)}

\item{y2}{if (replace_NA=TRUE & is.na(y2)): 100}

\item{replace_NA}{should NA be replaced at x1,y1,x2,y2 ??}
}
\value{
switch(as.character(type),
                 prop = scale* (x+offset)/centervalue ,
                 negprop = 2*scale - scale*(x +offset)/centervalue ,
                 antiprop=scale* centervalue/(x+offset),
                 identity= x)
}
\description{
Convert Values of Indicators to utility-values.
}
