% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_input_functions.R
\name{validateConfig}
\alias{validateConfig}
\title{Test validity of mcdsuppportshiny configuration}
\usage{
validateConfig(configList, dtAlternativen)
}
\arguments{
\item{dtAlternativen}{}
}
\description{
Test validity of mcdsuppportshiny configuration
}
