% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_reactive_helper_functions.R
\name{calculatecenterfunc}
\alias{calculatecenterfunc}
\title{Title}
\usage{
calculatecenterfunc(centervalue, x, offset, trim = 0, na.rm = FALSE)
}
\arguments{
\item{na.rm}{}
}
\value{

}
\description{
Title
}
\examples{

}
