% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_input_functions.R
\name{getIndikatorensetting}
\alias{getIndikatorensetting}
\title{Get Settings of Indicators}
\usage{
getIndikatorensetting(x, positive_utility_settings = utility_settings(),
  negative_utility_settings = utility_settings(util_func = "negprop"),
  agg_func = "weighted.sum", include_parent = TRUE,
  open.maxdepth = Inf, standardweight = 30)
}
\arguments{
\item{x}{Configuration list, see \code{\link{validateConfig}}.}

\item{positive_utility_settings}{standard utitily function settings, \code{\link{utility_settings}()}.}

\item{negative_utility_settings}{standard utitily function settings for negative weights, \code{\link{utility_settings}(util_func="negprop")}.}

\item{agg_func}{standard aggregation function.}

\item{include_parent}{Should new settings apply also to parent itself, or only to children?}

\item{open.maxdepth}{}

\item{minweight}{only to determine, if negative settings should be included}
}
\description{
Get Settings of Indicators
}
