% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_input_functions.R
\name{rgetIndikatorensetting}
\alias{rgetIndikatorensetting}
\title{Internal version of getIndikatorensetting}
\usage{
rgetIndikatorensetting(x, depth = 0, parent = "Szenarioergebnis",
  util_func = "prop", util_mean = "mean", util_offset = ifelse(util_func
  == "antiprop", 10, 0), util_scale = 100, agg_func = "weighted.sum",
  include_parent = TRUE, open.maxdepth = Inf, color = NA,
  standardweight = 30)
}
\arguments{
\item{color}{}
}
\value{

}
\description{
Internal version of getIndikatorensetting
}
