% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_input_functions.R
\name{utility_settings}
\alias{utility_settings}
\title{Helper function. Standard configuration of utility function}
\usage{
utility_settings(util_func = "prop", util_mean = "mean",
  util_offset = ifelse(util_func == "antiprop", 10, 0), util_scale = 100)
}
\arguments{
\item{util_scale}{}
}
\value{
list of four:
           list(util_func=util_func,
                util_mean=util_mean,
                util_offset=util_offset,
               util_scale=util_scale)
}
\description{
Helper function. Standard configuration of utility function
}
