% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_reactive_helper_functions.R
\name{utilityfunc}
\alias{utilityfunc}
\title{Convert Values of Indicators to utility-values.}
\usage{
utilityfunc(x, type, offset = ifelse(type == "antiprop", 50, 0),
  centervalue = calculatecenterfunc("mean", x, offset), scale = 100)
}
\arguments{
\item{x}{Numeric. Value to be converted.}

\item{type}{One of c("prop", "negprop", antiprop", "identity")}

\item{offset}{Offst to be added to x.}

\item{centervalue}{Numeric. Mean of utility values.}

\item{scale}{Scale.}
}
\value{
switch(as.character(type),
                 prop = scale* (x+offset)/centervalue ,
                 negprop = 2*scale - scale*(x +offset)/centervalue ,
                 antiprop=scale* centervalue/(x+offset),
                 identity= x)
}
\description{
Convert Values of Indicators to utility-values.
}
